Add a new ebuild that installs the fakemodem dbus config file, and use it.
The config file is necessary for tests to run the fakemodem program,
which is controlled over dbus, and it's much easier to just have it
present on the target system than to install it to /etc as part of the
deps mechanism within autotest.

The autotest-deps ebuild can't do this itself due to the
autotest-deponly style.

BUG=None
TEST=Run network_ModemManagerSMS test

Change-Id: Idca22363b2c22afedda57f583ff58eee0d26c568
Reviewed-on: https://gerrit.chromium.org/gerrit/11863
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Tested-by: Nathan J. Williams <njw@chromium.org>
Commit-Ready: Nathan J. Williams <njw@chromium.org>