Added courgette ebuild.
Courgette is pulled from http://git.chromium.org/chromium/src/courgette.git at
a pinned version specified in the ebuild, and built as a library with the
necessary headers to use it exported. Well, probably not everything needed,
but it's a start.

This isn't anything depending on this ebuild yet, so it shouldn't have any
real effect on the build yet.

BUG=chromium-os:22738
TEST=Built after doing a setup board, all dependancies needed were pulled in.

Change-Id: I6839648c81db514dfe6864fcd3d0ce4d8e547882
Reviewed-on: https://gerrit.chromium.org/gerrit/12358
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>