Unconditionally enable VBOOT_DEBUG in u-boot on x86
On ARM platforms, VBOOT_DEBUG was disabled for non-cros_workon
builds because it slows down the boot time significantly. However,
this is not the case on x86.
On the downside, if VBOOT_DEBUG is not enabled, vboot's messages
won't end up in /sys/firmware/log.
Hence, enable VBOOT_DEBUG all the time on x86, not just when
cros_debug is set.

BUG=none
TEST=boot a release image for lumpy and see vboot messages appear
     in /sys/firmware/log.

Change-Id: Ief8f31cf36a9f5c9e641dd88ff4f4f733421fa02
Reviewed-on: https://gerrit.chromium.org/gerrit/12160
Commit-Ready: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Tested-by: Stefan Reinauer <reinauer@chromium.org>