Work around broken strip_tests functionality in 17.0.947.0.
The strip_tests feature doesn't work in chromeos-chrome-17.0.947.0,
so let's temporarily disable it to workaround these bugs.

I've intentionally not rev'd Chrome so that developers will continue
using the existing prebuilts until Chrome is fixed and rev'd. Once
the new version of Chrome is ready, this change will be automatically
reverted.

BUG=chromium-os:23420
TEST=Verify strip_tests is no longer passed to Chrome makefile
     in this version.

Change-Id: Id7309668bf813b5a38af521497a79d693385e5e2
Reviewed-on: https://gerrit.chromium.org/gerrit/12162
Reviewed-by: Ryan Cui <rcui@chromium.org>
Tested-by: David James <davidjames@chromium.org>