Insert qemu-kvm into dependencies for hard-host-depends.
The hard-host-depends package used to have a dependency on qemu-user, but that
functionality was recently replaced by an upgrade to qemu-kvm.  A blocking dependency
on qemu-user was inserted into hard-host-depends, but a regular dependency
on qemu-kvm is needed as a replacement.

BUG=None
TEST=`sudo emerge --unmerge qemu-kvm`
Verify that qemu-kvm is not installed.
`sudo emerge hard-host-depends`
Verify that qemu-kvm is installed.

Change-Id: I33852cb02c97115fab7bcd277514def6d7d00b63
Reviewed-on: https://gerrit.chromium.org/gerrit/12221
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Matt Tennant <mtennant@chromium.org>
Tested-by: Matt Tennant <mtennant@chromium.org>