gmock/dialog env: only set config scripts for target boards
We don't want to try and force the config script settings for the cros
sdk itself.  We aren't cross-compiling, so the default values work just
fine.  Otherwise, when we try building gmock, things fail to build as
it tries to use a script that doesn't exist and can't find gtest.

TEST=`emerge gmock dialog` works again
TEST=`emerge-arm-generic gmock dialog` still works

Change-Id: Ia41c2edc133a9e9ff64e305d342fc0cd93793140
Reviewed-on: https://gerrit.chromium.org/gerrit/12332
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>