u-boot: Add config and fdt to USE_EXPAND list
This commit changes u-boot ebuild to acquire configurations from USE
flags, which is a preferred way of dealing with increasing irregularity
of naming conventions.

The boards in the snippet below are those have a valid U-Boot config and
device tree available.  I hope these are also the only boards that we
build U-Boot for, and for the rest of other boards, we should have set
dependency right so that U-Boot is not built when we build_packages for
those boards.

BUG=chromium-os:23869
TEST=Run script snippet below successfully

>>>>
for BOARD in lumpy \
             stumpy \
             tegra2_aebl \
             tegra2_arthur \
             tegra2_asymptote \
             tegra2_dev-board \
             tegra2_dev-board-opengl \
             tegra2_kaen \
             tegra2_seaboard \
             tegra2_wario \
             waluigi \
             x86-alex \
             x86-alex_he \
             x86-mario \
             x86-mario64
do
  emerge-${BOARD} chromeos-u-boot
done
<<<<

Change-Id: I6dae731837b4f9e74e9fba12c9c574b9083f4ed0
Reviewed-on: https://gerrit.chromium.org/gerrit/12486
Commit-Ready: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>