Added clang invocation to the sysroot_wrapper.
The main purpose of this CL is to enable clang warnings transparently for
any packages that need it, via a -clang options passed to the compiler. Both
clang and gcc warnings will be emitted for now.

The default behavior is the same.

BUG=none
TEST=rebuild chromeos with x86-zgb target. Ran suite_Smoke and bvt autotests.
Also tested the wrapper build time overhead. Without -clang it is almost the same.

Change-Id: I8209cc600bb60fc03c4fe7fa5bfdae8f348dbb64
Reviewed-on: https://gerrit.chromium.org/gerrit/12714
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: asharif <asharif@chromium.org>
Tested-by: Bhaskar Janakiraman <bjanakiraman@chromium.org>
Commit-Ready: Bhaskar Janakiraman <bjanakiraman@chromium.org>