Add a tool to provide command line access to device policies.
This CL adds the ebuild changes. This version of the CL installs a second
copy of the proto files to the /usr/share/protofiles location.

Subsequent CLs will update all projects that need proto files to use this
copy as well and then we can get rid of the old location that is
/usr/include/proto and is not being installed in the images.

BUG=chromium-os:22168
TEST=The policy_reader should be properly built and installed.

Change-Id: I87ac97f2909acad5b2991988e496c72b0ad7bec0
Reviewed-on: https://gerrit.chromium.org/gerrit/10780
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Commit-Ready: Julian Pastarmov <pastarmovj@chromium.org>
Tested-by: Julian Pastarmov <pastarmovj@chromium.org>