GD: drop local/unused/broken perl module
I can't find any packages actually using this, and it doesn't cross
compile at all.  While it seems like it builds for x86, it actually
uses the host compiler and so produces 64bit objects.  Nor has it
been keyworded for ARM.

Since the only target that this works for is the host, the existing
GD ebuild in the portage/ base should work just as well for that.
So drop our local copy.

BUG=chromium-os:24039
TEST=None

Change-Id: I957347d6099aaf4fa04ca2a3a0c21ad4aa751b59
Reviewed-on: https://gerrit.chromium.org/gerrit/12993
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>