gdbm: drop local copy
We have a local ebuild just to disable the compat library (due to relink
failures with the compat library when installing).  Newer versions from
upstream have a USE flag to disable the compat library, so we should be
able to switch easily to the upstream ebuild.

BUG=chromium-os:24038
TEST=`emerge gdbm perl ruby` still works
TEST=`emerge grub` still works (only package that uses ruby)
TEST=`emerge $(qlist -I dev-perl)` still works (all the local perl modules)
TEST=`emerge-amd64-generic gdbm` still works
TEST=`emerge-arm-generic gdbm` still works
TEST=`emerge-x86-generic gdbm` still works
CQ-DEPENDS=13009

Change-Id: Iba0c794634696c7c5e4dcc9b1828186ca5de419e
Reviewed-on: https://gerrit.chromium.org/gerrit/13010
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>