Add a USE flag to enable pageflipping.
Page flipping was disabled due to stability problems observed on
some platforms. The USE flag will make it easier to test and
(hopefully) track down the root of the issue.

Reordered the noflips and display-order patches. This results in
the line numbers being slightly wrong, but the two changes are
far enough apart they should continue to be resolved correctly.

BUG=chromium-os:15675
TEST=Manually set USE=pageflipping. X log now indicates that it
  is using page flipping.

Change-Id: I7f898ac01b2a1e5dce2518c3dde55cc765ae36f9
Reviewed-on: https://gerrit.chromium.org/gerrit/13095
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>