chromeos-base: install direct editor/pager wrappers
On the target hardware, we don't need the editor/pager shell scripts
to try and figure things out for us.  Instead, symlink to the only
values we have available.  In the cros-sdk, we still want these as
developers can change the settings based on their own tastes.

BUG=None
TEST=`emerge chromeos-base` does not install any symlinks
TEST=`emerge-x86-generic chromeos-bas` installs symlinks

Change-Id: I18a4ae0d1e8a37559fe1e5621fadadf52328bb76
Reviewed-on: https://gerrit.chromium.org/gerrit/13435
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>