integrate in known good manifest data
Being that people are finally on vacation and the tree is quiet,
integrate in maniest data- this touches majority of the tree, and
since we still have trivial merging forced, this needs to land during
quiet.

Specifically, pull it from upstream, and from our portage snapshot.
This is accomplished via `pinspect digests chromiumos` and a fair
bit of shell glue combined w/ pmaint digest and custom python.

This brings the versions missing data down to 17.4% of chromiumos,
104 packages out of 602 that lack data.

BUG=chromium-os:24028
TEST=normal emerge interactions

Change-Id: Ib1e14ce1e08362a3e41091753fef218b1e193bdf
Reviewed-on: https://gerrit.chromium.org/gerrit/13483
Commit-Ready: Brian Harring <ferringb@chromium.org>
Reviewed-by: Brian Harring <ferringb@chromium.org>
Tested-by: Brian Harring <ferringb@chromium.org>