utouch-evemu: add slot0 event when playing back a device event file
When playing back a device event file, check if slot0 exists
in the first packet. If it does not exist, insert a slot0 event
with a proper timestamp. This behavior can be switched on/off
by a command option (--insert-slot0).

BUG=chromium-os:24653
TEST=Use evemu-play to play back a scroll device event file.
The scroll should always work.

Change-Id: I3f250b09ab8ce388733e92b37d3fec9377316f3c
Reviewed-on: https://gerrit.chromium.org/gerrit/13558
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>