cros-workon: make eclass support any number of project entries, take 2
This generalisation is fully backwards compatible and for all
existing ebuilds, as bash treats scalar variables as one-item
arrays.
Even for src_info, all code that runs that function and calls eval on
the result will remain functioning, as arrays can be expanded as
regular variables in which case only the first item is printed.

This originally broke the tree on mesa, because the new eclass changes
DESTDIR behaviour from implicit (${S} dereferenced on copy) to explicit
(CROS_WORKON_DESTDIR which dereferences ${S} on eclass inherit). Fixed
the ebuild in the same commit.

BUG=chromium-os:25338
TEST=trybot x86-generic (nothing breaks)
TEST=build it with experimental multi-project ebuild, as no current
projects use it

Change-Id: I77b3bdebaff3f60c60e2a7bfba46623ee6007e9e
Reviewed-on: https://gerrit.chromium.org/gerrit/25616
Commit-Ready: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>