cros-workon: make eclass support any number of project entries
This generalisation is fully backwards compatible and for all
existing ebuilds, as bash treats scalar variables as one-item
arrays.
Even for src_info, all code that runs that function and calls eval on
the result will remain functioning, as arrays can be expanded as
regular variables in which case only the first item is printed.

BUG=chromium-os:25338
TEST=trybot x86-generic (nothing breaks)
TEST=build it with experimental multi-project ebuild, as no current
projects use it

Change-Id: I449ccbbca3cf6df625acafcb0b725c919a7e78fb
Reviewed-on: https://gerrit.chromium.org/gerrit/23908
Commit-Ready: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>