cros-workon: fix vcsid propagation
Observations:
1) Very few (~4) ebuilds actually use VCSID define for anything
2) Even less make use of it as "-DVCSID=" C-style definition
3) Escaping the quotes cannot be done universally, courtesy of libtool,
and thanks to the fact that the definition is dropped into C sources
as-is, proper string quoting is expected to come from outside.
4) Some packages fail with the original VCSID definition and already
explicitly filter it as a consequence.

This moves the VCSID definition into a passive variable export where the
ebuild is unaffected unless it wants to be. All affected ebuilds are fixed
for the "new way".

BUG=chromium-os:24819
TEST=emerge-x86-generic openssl cromo gobi-cromo-plugin power_manager
flimflam gestures

Change-Id: I207fe70847821ce941af47d693b7299734f03c04
Reviewed-on: https://gerrit.chromium.org/gerrit/13807
Commit-Ready: Zdenek Behan <zbehan@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>