cros-workon: fix vcsid propagation, take 2
Observations:
1) Very few (~4) ebuilds actually use VCSID define for anything
2) Even less make use of it as "-DVCSID=" C-style definition
3) Escaping the quotes cannot be done universally, courtesy of libtool,
and thanks to the fact that the definition is dropped into C sources
as-is, proper string quoting is expected to come from outside.
4) Some packages fail with the original VCSID definition and already
explicitly filter it as a consequence.

!5) Some CFLAGS that appear after the -DVCSID= statement may be
disabled, due to the oddities with string quoting. This can hide bugs or
create bugs in code that does not use VCSID at all.

This moves the VCSID definition into a passive variable export where the
ebuild is unaffected unless it wants to be. All affected ebuilds are fixed
for the "new way".

BUG=chromium-os:24819
TEST=below
1) emerge-x86-generic openssl cromo gobi-cromo-plugin power_manager flimflam gestures
2) trybot x86-generic

Change-Id: Ia5189989a0af67105d1fa5876fc7d08aa449fc96
Reviewed-on: https://gerrit.chromium.org/gerrit/14194
Tested-by: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Zdenek Behan <zbehan@chromium.org>