x11-base/xorg-server-1.9.3: Add Chromium Mouse Accel Profile
This adds a new acceleration profile type to xorg-server. It is based
on the algorithm we used for our touchpad work. In short, it is a
parabola at the low end, and a straight line at the high end.

The details of the algorithm are largely detailed in a comment in the
new patch.

BUG=chromium-os:24995
TEST=tested with many mice on a device

Change-Id: Icc2fa63c86c1d12640e882797500d234b48522a5
Reviewed-on: https://gerrit.chromium.org/gerrit/14065
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>