x11-drivers/xf86-input-evdev: Allow swapping left/right mouse buttons.
X can already allow arbitrary buttons to be swapped, but it happens at
a level where we can't distinguish between touchpads and mice. When a
user wants to swap left/right buttons, we want to do it only for mice,
so we edit the mouse driver (in this patch) to have the option to swap
buttons.

BUG=chromium-os:24996
TEST=tried swapping buttons many times on a device

Change-Id: Iff5e125f79ae06434a39097f48aa150f3b789b0c
Reviewed-on: https://gerrit.chromium.org/gerrit/14068
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>