Unmask the media-libs/jpeg to version 8b
I need to install libv4l to enable video capture from webcams for
OpenCV. The libv4l requires 8b verison of jpeg and so I relax the
mask from 6b to 8b. I find the 6b version is more than 3 years old
so I think the upgrade is reasonable.

I realize the jpeg library is used by many people so I have tested
it extensively below.

We also need to force chromeos-chrome to version bump as it uses
the system libjpeg, and we need it to rebuild to get the new SONAME.

BUG=chrome-os-partner:6497,chromium-os:24042
TEST='emerge-x86-alex media-libs/jpeg' works
TEST='emerge-x86-alex media-libs/tiff' works
TEST='emerge-x86-alex x11-libs/gtk+' works
TEST='emerge-x86-alex chromeos-base/chromeos-chrome' works
TEST='emerge-x86-zgb media-libs/jpeg' works
TEST='emerge-x86-zgb x11-libs/gtk+' works
TEST='emerge-amd64-generic media-libs/jpeg' works
TEST='emerge-amd64-generic x11-libs/gtk+' works
TEST='emerge-amd64-generic chromeos-base/chromeos-chrome' works
TEST='emerge-tegra2_kaen media-libs/jpeg' works
TEST='emerge-tegra2_kaen media-libs/tiff' works
TEST='emerge-tegra2_kaen x11-libs/gtk+' works
TEST='emerge-tegra2_kaen chromeos-base/chromeos-chrome' works
TEST=Web browsing with Chrome on pages with JPEG images works

Change-Id: I766fb61883a9ff9eaeba2feba6c9443a17b01944
Reviewed-on: https://gerrit.chromium.org/gerrit/14145
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>