libglade: fix building with old pango but new glib
The current version of libglade that we're building enables the
G_DISABLE_DEPRECATED macro because it doesn't need it, but the
older version of pango that we're using does.  So libglade disables
the deprecated stuff, includes pango which needs it, and then we
get errors from the pango headers.  Build libglade for now w/out
that macro.

Also, the libglade pkg-config.m4 code is old and searches $PATH
for "pkg-config" rather than for for "$CHOST-pkg-config".  Force
the configure test to go the way we need so we don't include and
link against files in /usr.

BUG=None
TEST=`emerge-tegra2 libglade` now passes

Change-Id: I06205394d9bccdf1181f885fe6e6bfb02d327ffc
Reviewed-on: https://gerrit.chromium.org/gerrit/14203
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>