Remove serial8250 kernel fragment.
This was originally added here:
  https://gerrit.chromium.org/gerrit/6661

It was intended only for the Tegra platform, but it was being enabled
on all platforms.

The proper way to do this is to add the config in the ARM or Tegra
splitconfig in the kernel.  ...and it looks like that's already been
done.  See:
  http://git.chromium.org/gitweb/?p=chromiumos/third_party/kernel.git;a=blob;f=chromeos/config/armel/common.config;h=67dbf8044af1eccb9d087dc09b1e833d38e4679b;hb=chromeos-3.0#l313

NOTE that this change will have the effect of _disabling_ serial8250
for x86 builds.  If you need serial8250 on x86 builds, you should
be building with the pcserial flag.

This cleanup is being done as part of the move of kernel ebuild
code to the eclass.

BUG=chromium-os:24521
TEST=Built and booted tegra2_kaen
TEST=Built and booted x86-generic

Change-Id: Ic38d70d403dbd8ec7d85f88584fcf7f29160a3cd
Reviewed-on: https://gerrit.chromium.org/gerrit/14569
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>