bluez: auto-generate PINs for keyboards
As recommended by the HID profile spec, rather than requiring that the
user enter a PIN on both the host and the pairing keyboard, generate a
PIN automatically and display it on the host for typing on the pairing
keyboard.

Patch sent upstream.

BUG=chromium-os:25212
TEST=tested pairing with available keyboard devices with both legacy and updated agents

Change-Id: I9b290d8e19dd10d47c55e58fa5f363891ff0de30
Reviewed-on: https://gerrit.chromium.org/gerrit/14652
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>
Commit-Ready: Scott James Remnant <keybuk@chromium.org>