bluez: change initial radio state to off
The bluetooth radio is off during firmware and kernel boot, but on
discovering a new adapter, bluetoothd will turn the radio on for
that adapter. This is both a security issue and a battery life issue.

Change bluetoothd's configuration so that the initial radio state is
off; once turned on by the user, bluetoothd remembers this setting
and will power on the radio on boot until the user turns it off again.

BUG=chromium-os:25475
TEST=verified radio is off with this patch, and stays on once enabled

Change-Id: Ice9654ffe716db5f1e174e093265a53e938e06b0
Reviewed-on: https://gerrit.chromium.org/gerrit/14795
Commit-Ready: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>