xf86-input-evdev: Expose devnode via read-only property
We currently use the exposed 'phys' along with udev to get the device
node. That's a lot of work, and difficult to do correctly with our
current version of udev. By exporting the device node itself, we can
migrate to that and stop using the 'phys' patch.

This patch is a rebased version of the patch at
http://lists.x.org/archives/xorg-devel/2011-February/018888.html
originaly signed off by Peter Hutterer.

BUG=chromium-os:25516
TEST=tested on device

Change-Id: Ic16dcb6a4d13679de4a19922e199ed99b96b78e4
Reviewed-on: https://gerrit.chromium.org/gerrit/14918
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>