Delete old cros-kernel eclass
Everyone should now be transitioned over to cros-kernel2, so get rid
of the old cros-kernel to avoid confusion.

If we ever feel like doing extra cleanup, we can always rename
cros-kernel2 back to cros-kernel.

This is part of the plan described in the refactor of the kernel
ebuilds.  See <https://gerrit.chromium.org/gerrit/14525>.  Originally
we were going to do the rename, but it doesn't seem worth it at this
point in time.

BUG=chromium-os:24521
TEST=Can still build packages.

Change-Id: I14d4ea26634247faa972d947eb6bb8483ca2478d
Reviewed-on: https://gerrit.chromium.org/gerrit/15251
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>