Fix minor issues in bashrc/bash_profile creation.
Specifically, for our cd invocation, do not assume we're invoked from
$HOME.  Some of the more special cros_sdk bits in the pipeline
may not play nice with this, and it's better to be absolute here.

Additionally, bash completion is only useful for interactive- thus
set it up in .bashrc which is sourced for interactive only.

A chroot upgrade script isn't currently required, although may
be forced depending on I28fedc64; if that CL requires it, an
upgrade script will be added.

BUG=None
TEST=cros_sdk --replace bash --norc; note lack of --board completion
TEST=cros_sdk; note --board completion still works.

Change-Id: I2fb4832a49e78049f0bfe8d7e198063f136a83cf
Reviewed-on: https://gerrit.chromium.org/gerrit/15297
Commit-Ready: Brian Harring <ferringb@chromium.org>
Reviewed-by: Brian Harring <ferringb@chromium.org>
Tested-by: Brian Harring <ferringb@chromium.org>