libchrome: install all available headers
Upgrading versions is painful as we hand maintain a list of headers that
base installs and as things move around upstream (deletions, additions,
moves, etc...), it's hard to track that.  So just have the ebuild install
all headers that exist.  This shouldn't be a problem as we install into
a base-specific dir, and if the headers aren't used, we don't actually
include them.

BUG=chromium-os:16623
TEST=build_packages for x86-alex still works

Change-Id: I9399dfcb2e722ee2e8decd0258984f8ba196c0ad
Reviewed-on: https://gerrit.chromium.org/gerrit/15416
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>