Add autotest-tests dependency on dev-python/numpy
It would be convenient to use numpy in a few autotests.  We already
install it on test builds, so this isn't adding any new code to the
test build, but it is adding one more user.

On my z600, emerge-x86-alex dev-python/numpy takes 101 seconds.

BUG=chromium-os:25423
TEST=emerge-x86-alex autotest-tests

Change-Id: I333111977786c3b912ae6bd97f7933697f203e6e
Reviewed-on: https://gerrit.chromium.org/gerrit/15677
Commit-Ready: David Rochberg <rochberg@chromium.org>
Reviewed-by: David Rochberg <rochberg@chromium.org>
Tested-by: David Rochberg <rochberg@chromium.org>