auto update: for 32bit/64bit boards, add USE=32bit_au
We want to build the auto update code as static 32bit binaries for
the boards that we are upgrading to 64bit in the field.  This is
plumbing to make that simpler in ebuilds.

Now boards can add USE=32bit_au to their make.conf and get this.

BUG=chromium-os:25370
TEST=`emerge libbb chromeos-installer` still work like normal
TEST=`emerge-arm-generic libbb chromeos-installer` still work like normal
TEST=`emerge-x86-mario libbb chromeos-installer` installs 32bit binaries (in an otherwise 64bit system)

Change-Id: I5710f4d1008f7ada259b2eeca9488686a8daecff
CQ-DEPEND=Ic2ca1aeeefa5ea2207d3c963ac856909cb40cae5
Reviewed-on: https://gerrit.chromium.org/gerrit/15783
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>