Install library version of verity with header files
This installs the static version of the verity library
along with the header files needed to use it.  This
happens to include pretty much all of the files in
the include directory because the library interface
headers rely on all the interdependent set of headers
in there.

BUG=chromium-os:25377
TEST=`emerge-amd64-generic verity` installs libdm-bht.a and puts headers
	into usr/include/verity
CQ-DEPEND=I4037c68d3fc2c1dfdcfed4a603b49b57dd06562b

Change-Id: Ie61befdab79529c975febca464f11a0293200855
Reviewed-on: https://gerrit.chromium.org/gerrit/15840
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>