Build and install the static verision of verity
This allows the static installer to do a simple call to the static
version of verity for the 32 to 64bit transition.  Note that
we still need a CL to force a 32bit build, but this part is
necessary regardless.

BUG=chromium-os:25377
TEST=/bin/verity-static gets installed
CQ-DEPEND=I93d542c5da96f3f30a1ce479f9cfc993e4e888ca

Change-Id: Ib7411decd0ec9d540c9ac7d13e5b36db7fc6a685
Reviewed-on: https://gerrit.chromium.org/gerrit/15843
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>