Add support for Linux KVM in the new eclass scheme.
An earlier CL enabled building kvm modules into the kernel.
About the time the CL was approved the scheme changed.
This new scheme has been used and tested on a platform.
We've also added in virtio consoles.

BUG=chrome-os-partner:7676
TEST=build a kernel with USE=kvm as an option.

Reboot.
modprobe kvm_intel
should work to the extent that the module is found. Even if vmx is not
enabled on the platform, the module will exist but might fail to load
if vmx is not turned on.

e.g.
localhost ~ # modprobe kvm_intel
localhost ~ #

Change-Id: I3071e432b1e288570065b9c07f4cdf5be1342d98
Reviewed-on: https://gerrit.chromium.org/gerrit/16080
Commit-Ready: Ronald G. Minnich <rminnich@chromium.org>
Tested-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>