Symlink verity-static to verity
This is in anticipation of having verity being built as 32bit
on 64bit targets.  We will have a dynamically linked 32bit binary
on a 64bit system without any of the necessary libraries to run it.
So, this simplifies things and just makes verity a symlink to
verity-static, and nobody needs to be any wiser.

BUG=chromium-os:25377
TEST=`emerge-amd64-generic verity` ensure there is a symlink in
   /build/amd64-generic/bin verity -> verity-static

Change-Id: I583b4a0d19f59aa5c34a7583e8d6a60d5f6e2c79
Reviewed-on: https://gerrit.chromium.org/gerrit/16118
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>