chromeos-bootimage: Don't set silent_console on x86
On x86, u-boot does not print its messages to the display but
to a memory console, so it's much faster and not visible to the user.
With silent_console set to 1, u-boot's log messages do not end up in
the memory log (/sys/firmware/log). This CL prevents that from
happening, making automated firmware builds from the firmware branch
contain u-boot's log messages.

This commit has no direct effect on production firmware images, it
only keeps ToT in sync with the described behavior so that we don't
have to fix this next time we do a firmware branch.

BUG=chrome-os-partner:7992
TEST=boot coreboot+u-boot release on stumpy/lumpy and see all of
     u-boot's messages in /sys/firmware/log

Change-Id: I2755355ef7e3852dfbb6ad424204ec6638502235
Reviewed-on: https://gerrit.chromium.org/gerrit/16143
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>