Include the Chrome OS fonts in hard-host-depends.
Recovery needs to generate internationalized messages for multiple
locales.  Because the message generation happens in the ebuild, we
need the Chrome OS fonts merged on the host, and not just on the
target.

This change also includes minor cleanup to remove a dependency that
appeared twice in the chromeos-base/chromeos ebuild.

BUG=chromium-os:26757
TEST=Boot a unit configured for Korean.
TEST=Generate and inspect recovery images for multiple locales.

Change-Id: I6a206eaafaf9fab572381e52799c8be257d833af
Reviewed-on: https://gerrit.chromium.org/gerrit/16447
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>