chromeos-base: Add cras user for the new audio server.
The audio server will be allowed access to the audio hardware.  The
system user will be allowed to use the server to play sounds.  Add the
cras user and make him a member of the audio group.  Add the system
user to the cras group.  Until the transition to using cras from
chrome is complete, leave chronos is the audio group.

BUG=none
TEST=security_AccountsBaseline(updated) and play youtube on mario.

Change-Id: Ia845fdf689db4db9a0e46db83727236582a423c4
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/16609
Reviewed-by: Mike Frysinger <vapier@chromium.org>