libchrome: drop building of trace_event
This object has clearly never been used because it relies on an external
undefined symbol base::GetCurrentProcId() which is in yet another file
(process_util_posix.cc) that isn't compiled.  Attempts to use it result
in linking errors.  So drop it from the build.

BUG=chromium-os:16623
TEST=`cbuildbot arm-generic-full` still works
TEST=`emerge metrics chromeos-imageburner chromeos-cryptohome chaps cromo cros-disks crash-reporter vpn-manager speech_synthesis` still works

Change-Id: I793b81829ef830ad8ee3ff840a696a61f3b22d18
Reviewed-on: https://gerrit.chromium.org/gerrit/16516
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>