Selectively build chromeos-bootimage when building chromeos
The chromeos-bootimage package is not a dependency for chromeos and as
such is not built when chromeos is emerged. This in turn means that
chromeos-bootimage breakages go unnoticed by the buildbots.

The suggested solution is to make chromeos-bootimage package a DEPEND
dependency for chromeos, i.e. the chromeos-bootimage package would be
built but not installed when chromeos package is emerged.

Actual enabling of this use flag will happen later on per board basis
through the respective make.conf files.

BUG=chromium-os:26696
TEST=manual

 . run USE=bootimage ./build_packages --board=link --oldchromebinary
 . observe new firmware generated in /build/link/firmware

Change-Id: I324cd350e4ab845b9ba6d32c6477025545892817
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/16558