libchromeos: do not mess with env in src_test
The env setup in src_compile propagates over to src_test automatically
(this behavior is required in ebuilds), so there's no need to try and
set it up again.  Further, this fixes double compiles of many files as
scons goes "oh, the CFLAGS changed, so i need to rebuild all the .o's"
which we don't want.

BUG=chromium-os:26736
TEST=`FEATURES=test emerge libchromeos` works

Change-Id: I7361fd1c657a5e5b55527505e08a551fc2de07a0
Reviewed-on: https://gerrit.chromium.org/gerrit/17015
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>