Pass Aura use flag as define to allow screen blank fix
Code was committed in third_party/chrontel to fix the screen blanking
on resize when Aura is in use. This is enabled by a define. This ebuild
change causes USE_AURA to be defined on the compiler command line if the
USE=aura flag is set for the ebuild.

BUG=chrome-os-partner:7347
TEST=do build and see -DUSE_AURA on compile command line

Change-Id: I7c3c9a2efec31c96ac47eead48b7c31e0921c24a
Reviewed-on: https://gerrit.chromium.org/gerrit/16789
Tested-by: Mark Hayter <mdhayter@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Mark Hayter <mdhayter@chromium.org>