cros-workon: introduce CROS_WORKON_SUBDIRS_TO_COPY variable
This makes sure that only certain subtrees are copied in 9999 builds.
Its primary use target is autotest-*, as it has multiple ebuilds that
build mutually disjoint parts of the source tree.
It can also be used by CQ/PFQ to recognise which changes are going to
affect which ebuilds.

BUG=chromium-os:26985
TEST=add a testing CROS_WORKON_SUBDIRS_TO_COPY statement into autotest-tests

Change-Id: I3cc3f58b4ad046de57c1e3a7e71a030b79d1a015
Reviewed-on: https://gerrit.chromium.org/gerrit/16932
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Zdenek Behan <zbehan@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>