daisy: Assign a correct HOSTCC in chromeos-u-boot ebuild.
Some utilities in u-boot tools directory should be compiled using a host CC.
Assign a correct one in order not to break the build.

BUG=chrome-os-partner:8246
TEST=run: emerge-daisy chromeos-u-boot
It is able to compile u-boot but failed in copying DTS files.
This error will be fixed in a later change.

Change-Id: I3bfab50f7021510080b8b99643c2fbd463fe9ea2
Reviewed-on: https://gerrit.chromium.org/gerrit/16976
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>