gcc: automatically execute ccache
For target boards, automatically execute ccache so that repetitive builds
are much faster.  A simple test shows that this is an improvement in the
workflow for package developers:
	$ time emerge-x86-alex cros-disks
	real    0m14.994s
	user    1m19.020s
	sys     0m9.860s

	$ time emerge-x86-alex cros-disks
	real    0m9.073s
	user    0m7.020s
	sys     0m3.200s

A new -noccache flag is added in case someone wants to bypass the support
in the wrapper.

While gcc is very picky about its argv[0] being the correct full path,
I don't think there should be issues with the basename(argv[0]) keeping
the ".real" suffix.  Tests so far show that it's OK.

Tests with full bots and cold caches show marginal overhead (i.e. all
cache misses).  So the penalty here should be acceptable considering
the significant gains we see when there are cache hits.

BUG=chromium-os:25613
TEST=See gerrit for discussion/more tests

Change-Id: Iad0e0e160a98cd92c707a24b20414b11c62b6e31
Reviewed-on: https://gerrit.chromium.org/gerrit/17187
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>