For Japanese, select the Motoya fonts for internal builds.
When chromeos-base/chromeos-fonts is built for the host in a chroot
with the correct private overlay, we want the Motoya fonts to be
available, since these are the preferred fonts for Japanese.

The Motoya fonts are only available in the case of a Chrome (not
Chromium) OS build, so when the private overlay is not available,
the package provides the alternative IPA fonts instead.

BUG=chromium-os:26757
TEST=in chroot:  ls /usr/share/fonts, and see the proper content depending on public/private overlay
TEST=build and inspect recovery message images from public and private overlays
CQ-DEPEND=I4a553703b0e14e33683734b895181a2bad5da2a4

Change-Id: I32c4561c47752d96d2769c5429c5e44fe630f8f9
Reviewed-on: https://gerrit.chromium.org/gerrit/17373
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Takayoshi Kochi <kochi@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>