libchrome: update to svn rev 125070
This updates us to the version of base we're using in Chrome 19.0.1061.0.
This also cuts us over to using aura (so no GTK, just X11).

Nothing currently uses this directly, but I've tested a few packages
against it, and they seem to be OK.

BUG=chromium-os:25872
TEST=`cros_run_unit_tests -p libchromeos` passes

Change-Id: Iacf87fc2d7f7004de7378d68f4b5db102667aa3d
Reviewed-on: https://gerrit.chromium.org/gerrit/17527
Reviewed-by: Eric Shienbrood <ers@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>