vboot_reference: clean up & convert to SLOT-ed libbase
Fix the ordering of variables: EAPI/cros-workon, then inherit, then
the rest.

Update to EAPI 4 for fun.  Drop all the dies that are now redundant.
Delete old bitmap handling logic that's been broken for a while, but
is only now fatal (due to the EAPI 4 change).

Include cros-debug because we depend on libchrome.  Then fix the
libchrome dep to pull in the SLOT and have the USE dep (just like
all the other ebuilds).

Call cros-debug-add-NDEBUG now that we pull in the eclass for it,
and export PKG_CONFIG so the build can use it.

Fix whitespace in src_install -- tabs, not spaces.

BUG=chromium-os:16623
BUG=chromium-os:27457
TEST=`emerge-x86-alex vboot_reference` builds & links CgptManagerTests against newer libbase

Change-Id: I7d2dce71b889e573295233b68edd6675edbfbc5f
Reviewed-on: https://gerrit.chromium.org/gerrit/17534
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>