xorg-server: Avoid a crash with non-core inputs
One of the changes between xorg-server 1.9.3-e24 and 1.11.99.902
makes dix/getevents.c:positionSprite unsafe for devices that do
not send core events. In that case miPointerSetPosition returned
NULL, and later dereferences of the scr variable would segfault.

This patch modifies the code flow to avoid a crash, and follow
the behavior of 1.9.

BUG=chromium-os:27529
TEST=x86-alex with external touchscreen no longer crashes on input,
  and still behaves properly with the trackpad or a USB mouse.

Change-Id: Ia9ccc32ee91489a296def42455c2748ba05bc4e0
Reviewed-on: https://gerrit.chromium.org/gerrit/17590
Reviewed-by: Stuart Abercrombie <sabercrombie@chromium.org>
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>