chromeos-base/xorg-conf: Set Tap Minimum Pressure for IS Synaptics pads.
We generally rely on pressure calibration to unify the various
touchpads, but it's not perfect. In this case, the value we pick for
the minimum pressure required to trigger a tap has to be done
carefully, and unfortunately it varies on different pads. This CL sets
the value explicitely for the IS (Image Sensor) Synaptics pads.

BUG=chromium-os:27554
TEST=tested on device

Change-Id: Idbd416b03ebdd35155dab41c6db985660b0623e4
Reviewed-on: https://gerrit.chromium.org/gerrit/17650
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>